Fixed some MSVC warnings.

+++ modules/juce_audio_formats/format/juce_AudioFormatReaderSource.cpp
@@ -60,17 +60,17 @@ void AudioFormatReaderSource::getNextAudioBlock (const AudioSourceChannelInfo& i
             if (newEnd > newStart)
             {
                 reader->read (info.buffer, info.startSample,
-                              newEnd - newStart, newStart, true, true);
+                              (int) (newEnd - newStart), newStart, true, true);
             }
             else
             {
-                const int64 endSamps = reader->lengthInSamples - newStart;
+                const int endSamps = (int) (reader->lengthInSamples - newStart);
 
                 reader->read (info.buffer, info.startSample,
                               endSamps, newStart, true, true);
 
                 reader->read (info.buffer, info.startSample + endSamps,
-                              newEnd, 0, true, true);
+                              (int) newEnd, 0, true, true);
             }
 
             nextPlayPos = newEnd;
